refactor: router를 app으로 이름 및 위치 변경
